+++ lib/active_admin.rb
@@ -6,6 +6,7 @@ require 'ransack_ext'
 require 'bourbon'
 require 'kaminari'
 require 'formtastic'
+require 'formtastic_i18n'
 require 'sass-rails'
 require 'inherited_resources'
 require 'jquery-rails'
@@ -91,7 +92,7 @@ module ActiveAdmin
     #
     # @param [Block] block A block to call each time (before) AA loads resources
     def before_load(&block)
-      ActiveAdmin::Event.subscribe ActiveAdmin::Application::BeforeLoadEvent, &block
+      ActiveSupport::Notifications.subscribe ActiveAdmin::Application::BeforeLoadEvent, &ActiveAdmin::Event.wrap_block_for_active_support_notifications(block)
     end
 
     # A callback is triggered each time (after) Active Admin loads the configuration files. This
@@ -109,7 +110,7 @@ module ActiveAdmin
     #
     # @param [Block] block A block to call each time (after) AA loads resources
     def after_load(&block)
-      ActiveAdmin::Event.subscribe ActiveAdmin::Application::AfterLoadEvent, &block
+      ActiveSupport::Notifications.subscribe ActiveAdmin::Application::AfterLoadEvent, &ActiveAdmin::Event.wrap_block_for_active_support_notifications(block)
     end
 
   end