+++ lib/active_admin/resource_controller/streaming.rb
@@ -20,7 +20,12 @@ module ActiveAdmin
       def stream_resource(&block)
         headers['X-Accel-Buffering'] = 'no'
         headers['Cache-Control'] = 'no-cache'
-        self.response_body = Enumerator.new &block
+
+        if ActiveAdmin.application.disable_streaming_in.include? Rails.env
+          self.response_body = block['']
+        else
+          self.response_body = Enumerator.new &block
+        end
       end
 
       def csv_filename
@@ -28,6 +33,7 @@ module ActiveAdmin
       end
 
       def stream_csv
+        headers['Content-Type'] = 'text/csv; charset=utf-8' # In Rails 5 it's set to HTML??
         headers['Content-Disposition'] = %{attachment; filename="#{csv_filename}"}
         stream_resource &active_admin_config.csv_builder.method(:build).to_proc.curry[self]
       end