+++ lib/generators/active_admin/assets/assets_generator.rb
@@ -8,7 +8,7 @@ module ActiveAdmin
         generate "redactor:install"
         rake "db:migrate"
         template 'active_admin.js.coffee', 'app/assets/javascripts/active_admin.js.coffee'
-        template 'active_admin.css.scss',  'app/assets/stylesheets/active_admin.css.scss'
+        template "active_admin.scss", "app/assets/stylesheets/active_admin.scss"
       end
 
     end