+++ spec/unit/generators/install_spec.rb
@@ -3,20 +3,21 @@ require 'rails_helper'
 describe "AA installation" do
   context "should create" do
 
-    it "active_admin.css.scss" do
-      expect(File.exists?(Rails.root + "app/assets/stylesheets/active_admin.css.scss")).to be_truthy
+    it "active_admin.scss" do
+      path = Rails.root + "app/assets/stylesheets/active_admin.scss"
+      expect(File.exists? path).to eq true
     end
 
     it "active_admin.js.coffee" do
-      expect(File.exists?(Rails.root + "app/assets/javascripts/active_admin.js.coffee")).to be_truthy
+      expect(File.exists?(Rails.root + "app/assets/javascripts/active_admin.js.coffee")).to eq true
     end
 
     it "the dashboard" do
-      expect(File.exists?(Rails.root + "app/admin/dashboard.rb")).to be_truthy
+      expect(File.exists?(Rails.root + "app/admin/dashboard.rb")).to eq true
     end
 
     it "the initializer" do
-      expect(File.exists?(Rails.root + "config/initializers/active_admin.rb")).to be_truthy
+      expect(File.exists?(Rails.root + "config/initializers/active_admin.rb")).to eq true
     end
 
   end