+++ spec/unit/namespace/register_page_spec.rb
@@ -1,9 +1,7 @@
 require 'rails_helper'
 
 describe ActiveAdmin::Namespace, "registering a page" do
-
   let(:application){ ActiveAdmin::Application.new }
-
   let(:namespace){ ActiveAdmin::Namespace.new(application, :admin) }
   let(:menu){ namespace.fetch_menu(:default) }
 
@@ -17,23 +15,21 @@ describe ActiveAdmin::Namespace, "registering a page" do
     end
 
     it "should create a new controller in the default namespace" do
-      expect(defined?(Admin::StatusController)).to be_truthy
+      expect(defined?(Admin::StatusController)).to eq 'constant'
     end
 
     it "should create a menu item" do
       expect(menu["Status"]).to be_an_instance_of(ActiveAdmin::MenuItem)
     end
-  end # context "with no configuration"
+  end
 
   context "with a block configuration" do
     it "should be evaluated in the dsl" do
-      expect {
-        namespace.register_page "Status" do
-          raise "Hello World"
-        end
-      }.to raise_error
+      expect{ |block|
+        namespace.register_page "Status", &block
+      }.to yield_control
     end
-  end # context "with a block configuration"
+  end
 
   describe "adding to the menu" do
     describe "adding as a top level item" do
@@ -42,9 +38,9 @@ describe ActiveAdmin::Namespace, "registering a page" do
       end
 
       it "should add a new menu item" do
-        expect(menu['Status']).to_not be_nil
+        expect(menu['Status']).to_not eq nil
       end
-    end # describe "adding as a top level item"
+    end
 
     describe "adding as a child" do
       before do
@@ -52,13 +48,15 @@ describe ActiveAdmin::Namespace, "registering a page" do
           menu parent: 'Extra'
         end
       end
+
       it "should generate the parent menu item" do
-       expect( menu['Extra']).to_not be_nil
+       expect( menu['Extra']).to_not eq nil
       end
+
       it "should generate its own child item" do
-        expect(menu['Extra']['Status']).to_not be_nil
+        expect(menu['Extra']['Status']).to_not eq nil
       end
-    end # describe "adding as a child"
+    end
 
     describe "disabling the menu" do
       before do
@@ -66,9 +64,10 @@ describe ActiveAdmin::Namespace, "registering a page" do
           menu false
         end
       end
+
       it "should not create a menu item" do
-        expect(menu["Status"]).to be_nil
+        expect(menu["Status"]).to eq nil
       end
-    end # describe "disabling the menu"
-  end # describe "adding to the menu"
+    end
+  end
 end