+++ spec/unit/namespace/register_resource_spec.rb
@@ -4,41 +4,40 @@ require 'rails_helper'
 #       Specifically, the dashboard is already defined.
 
 describe ActiveAdmin::Namespace, "registering a resource" do
-
   let(:application){ ActiveAdmin::Application.new }
-
   let(:namespace){ ActiveAdmin::Namespace.new(application, :admin) }
-
   let(:menu){ namespace.fetch_menu(:default) }
 
   context "with no configuration" do
     before do
       namespace.register Category
     end
+
     it "should store the namespaced registered configuration" do
       expect(namespace.resources.keys).to include('Category')
     end
+
     it "should create a new controller in the default namespace" do
-      expect(defined?(Admin::CategoriesController)).to be_truthy
+      expect(defined?(Admin::CategoriesController)).to eq 'constant'
     end
+
     skip "should not create the dashboard controller" do
-      defined?(Admin::DashboardController).to_not be_truthy
+      defined?(Admin::DashboardController).to_not eq 'constant'
     end
+
     it "should create a menu item" do
       expect(menu["Categories"]).to be_a ActiveAdmin::MenuItem
       expect(menu["Categories"].instance_variable_get(:@url)).to be_a Proc
     end
-  end # context "with no configuration"
+  end
 
   context "with a block configuration" do
     it "should be evaluated in the dsl" do
-      expect {
-        namespace.register Category do
-          raise "Hello World"
-        end
-      }.to raise_error
+      expect{ |block|
+        namespace.register Category, &block
+      }.to yield_control
     end
-  end # context "with a block configuration"
+  end
 
   context "with a resource that's namespaced" do
     before do
@@ -50,7 +49,7 @@ describe ActiveAdmin::Namespace, "registering a resource" do
       expect(namespace.resources.keys).to include('Mock::Resource')
     end
     it "should create a new controller in the default namespace" do
-      expect(defined?(Admin::MockResourcesController)).to be_truthy
+      expect(defined?(Admin::MockResourcesController)).to eq 'constant'
     end
     it "should create a menu item" do
       expect(menu["Mock Resources"]).to be_an_instance_of(ActiveAdmin::MenuItem)
@@ -89,7 +88,7 @@ describe ActiveAdmin::Namespace, "registering a resource" do
         namespace.register Category
       end
       it "should add a new menu item" do
-        expect(menu['Categories']).to_not be_nil
+        expect(menu['Categories']).to_not eq nil
       end
     end # describe "adding as a top level item"
 
@@ -100,10 +99,10 @@ describe ActiveAdmin::Namespace, "registering a resource" do
         end
       end
       it "should generate the parent menu item" do
-        expect(menu['Blog']).to_not be_nil
+        expect(menu['Blog']).to_not eq nil
       end
       it "should generate its own child item" do
-        expect(menu['Blog']['Categories']).to_not be_nil
+        expect(menu['Blog']['Categories']).to_not eq nil
       end
     end # describe "adding as a child"
 
@@ -114,7 +113,7 @@ describe ActiveAdmin::Namespace, "registering a resource" do
         end
       end
       it "should not create a menu item" do
-        expect(menu["Categories"]).to be_nil
+        expect(menu["Categories"]).to eq nil
       end
     end # describe "disabling the menu"
 
@@ -126,7 +125,7 @@ describe ActiveAdmin::Namespace, "registering a resource" do
           end
         end
         it "should not show up in the menu" do
-          expect(menu["Posts"]).to be_nil
+          expect(menu["Posts"]).to eq nil
         end
       end
       context "when optional" do
@@ -136,7 +135,7 @@ describe ActiveAdmin::Namespace, "registering a resource" do
           end
         end
         it "should show up in the menu" do
-          expect(menu["Posts"]).to_not be_nil
+          expect(menu["Posts"]).to_not eq nil
         end
       end
     end
@@ -147,14 +146,14 @@ describe ActiveAdmin::Namespace, "registering a resource" do
       it "should be namespaced" do
         namespace = ActiveAdmin::Namespace.new(application, :one)
         namespace.register Category
-        expect(defined?(One::CategoriesController)).to be_truthy
+        expect(defined?(One::CategoriesController)).to eq 'constant'
       end
     end
     context "when not namespaced" do
       it "should not be namespaced" do
         namespace = ActiveAdmin::Namespace.new(application, :two)
         namespace.register Category
-        expect(defined?(Two::CategoriesController)).to be_truthy
+        expect(defined?(Two::CategoriesController)).to eq 'constant'
       end
     end
   end # describe "dashboard controller name"