+++ spec/unit/resource_spec.rb
@@ -35,7 +35,7 @@ module ActiveAdmin
 
     describe '#decorator_class' do
       it 'returns nil by default' do
-        expect(config.decorator_class).to be_nil
+        expect(config.decorator_class).to eq nil
       end
       context 'when a decorator is defined' do
         let(:resource) { namespace.register(Post) { decorate_with PostDecorator } }
@@ -80,9 +80,9 @@ module ActiveAdmin
     describe "#belongs_to" do
 
       it "should build a belongs to configuration" do
-        expect(config.belongs_to_config).to be_nil
+        expect(config.belongs_to_config).to eq nil
         config.belongs_to :posts
-        expect(config.belongs_to_config).to_not be_nil
+        expect(config.belongs_to_config).to_not eq nil
       end
 
       it "should set the target menu to the belongs to target" do
@@ -218,12 +218,12 @@ module ActiveAdmin
       context "when breadcrumb is set" do
         context "when set to true" do
           before { config.breadcrumb = true }
-          it { is_expected.to be_truthy }
+          it { is_expected.to eq true }
         end
 
         context "when set to false" do
           before { config.breadcrumb = false }
-          it { is_expected.to be_falsey }
+          it { is_expected.to eq false }
         end
       end
     end
@@ -304,22 +304,24 @@ module ActiveAdmin
           :before_filter, :skip_before_filter,
           :after_filter, :skip_after_filter,
           :around_filter, :skip_filter
-        ].each do |filter|
-          it "delegates #{filter}" do
-            expect(resource.send(filter)).to eq "called #{filter}"
+        ].each do |method|
+          it "delegates #{method}" do
+            expected = method.to_s.dup
+            expected.sub! 'filter', 'action' if ActiveAdmin::Dependency.rails >= 4
+            expect(resource.send(method)).to eq "called #{expected}"
           end
         end
       end
 
-      if Rails::VERSION::MAJOR == 4
+      if ActiveAdmin::Dependency.rails >= 4
         context "actions" do
           [
             :before_action, :skip_before_action,
             :after_action, :skip_after_action,
             :around_action, :skip_action
-          ].each do |action|
-            it "delegates #{action}" do
-              expect(resource.send(action)).to eq "called #{action}"
+          ].each do |method|
+            it "delegates #{method}" do
+              expect(resource.send(method)).to eq "called #{method}"
             end
           end
         end